+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Fixes
 
+- Fix an issue on macOS 13 that the stepper in the custom tab width view worked only once.
 - Fix an issue that Snippet settings pane was not localized in Italian.
 - Fix missing translation in Italian (Thanks to Roccobot!).